Add 'events' to resolve.fallback to fix build
This fixed it on my end. Here's the error for reference: 

```
Module not found: Error: Can't resolve 'events' in 'C:\Users\freshgum\Development\cinny\src\app\molecules\import-e2e-room-keys'

BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
	- add a fallback 'resolve.fallback: { "events": require.resolve("events/") }'
	- install 'events'
If you don't want to include a polyfill, you can use an empty module like this:
	resolve.fallback: { "events": false }
```