Add basic support for displaying emotes
This only handles *displaying* emote events in a specific manner
recognizable as an emote.

It requires passing the sender name to message content elements, so that
the name can be in-lined into the message.

Sending emotes requires more changes. Namely better support for
commands. It would be required that they're not tied to the
completion framework for activation, but hooked on message being
sent and starting with a `/`. (WIP elsewhere)