Do not send messages when confirming an input method selection
When using a CJK input method, the enter key is used to confirm the
current conversion, so it shouldn't send the message yet. Normally
this can be checked with the isComposing property of the native
event, but that doesn't work on Safari on macOS.

Instead, check for keyCode 229, which is what you get when confirming
conversions, and ignore it.